3321: Handle the submission of AuthnRequests
Will call out to the API and redirect to start on success.

As capybara doesn't like posting data directly and redirects following that we
have introduced a test page which will submit our SAML for us. This page should
not be visible outside of tests

Authors: @chrisholmes @thomasalanlee @vedavallikanala @vixus0 @richardtowers