3321: Use HTML validation attributes instead of JS configuration
jQuery validate allows you to declare which fields are required using
`required` attributes. This is more convenient than configuring the
rules in JS. It also allows you to specify the message in a `data-msg`
attribute, which is convenient because it allows us to translate the
message on the server using Rails' locales.

Authors: @oswaldquek @richardtowers