3320 Fixed back button on sign-in page to link back to start page.
We were using rails ':back' earlier, this caused problems when we
switched language on the page as rails back  would use referrer to redirect.
However, we will have to relook at this implementation when we play the
non-repudiation flow story as the context of back will change during this flow
and it will have to be confirm_your_identity page and not start page.

Authors: @veda @oswaldquek @vixus0 @lloydnye