3416: Add a JSON-format log
This adds a log in JSON format - currently we're just mirroring
everything in the standard log with the addition of the session id if
it's available.

It would be nice to record exceptions on one line - currently each line
of the backtrace gets a new line in the log file, and we'd rather the
backtrace was logged on the same line in its own field.

Authors: @oswaldquek @leoshaw