3345: Refactor for Federation Info
On the About and Sign In pages, we need to send a custom variable to
Piwik defining the relying party that issued the AuthnRequest. In old
frontend, this is done using the request issuer ID (also called
transaction entity ID) obtained from Policy. This value is obtained
along with the list of IDPs.

However, we were not sending the transaction entity ID along with the
list of IDPs from our API. This has now been fixed in ida-frontend. The
following changes allow new frontend to deal with the modified response
from the API and we can now send the appropriate reports to Piwik.

Authors: @richardtowers @vixus0