3503: send headers and cookies with request statement
Persistence appears to become broken when headers and cookies are not set on the
same line as the request. Therefore, we have moved to the same line.

There's an interesting quirk with cookies following this change. The client
library was ignoring the key fields in the cookie hash. To overcome this quickly
we have a workaround to transform the value fields into a cookie object that
contains a key and value.

Authors: @chrisholmes @davbo