bau: provide a 404 page matching live
This change provides a 404 page for the user that matches the behaviour of live.

In order to catch the "404" Exception I have had to configure the app to refer
to our own routes for errors where it will route to the new ErrorsController. We
may want to route the 500 error this way in the future.