BAU: standardise route URLs
This makes the URLs consistent between dev and non-dev environments, so
we use dashes as separators across the board.  The reason for this is
consistency: it doesn't matter which we use in the dev environment, but
having differences makes it easy to introduce routes that don't map
correctly to the old frontend.
This also fixes a route (forgot-company) that used the wrong separators
in non-dev.

Authors: @yolinas @leoshaw