bau: Small fixes recommended by rubocop
Running rubocop with a stricter config than the GDS defaults raises some
interesting things such as freezing strings which are intended to be
constant (strings are mutable by default in Ruby). I think we should
consider making some changes to our rubocop config to catch these
things.

Includes a separate change which will only include the JasmineRails
routes if we are in test or development env, they are currently excluded
by the dependency gem not being available in production.