3517: integrate raven into the rails logger
Now when we send error messages to the Rails logger they will be sent to Sentry
using Raven as well as our other loggers. This should allow us to not have to explicitly capture exceptions
where we are currently logging errors.

In order to integrate with the Rails logger we provider a logger for raven that
implements the ::Logger interface. It is then added to the Rails logger as an
additional output using ActiveSupport::Logger::broadcast.

Also, provided an integration test.