3491: Enabling javascript on the select phone page.
Added ids to the html form questions so that the javascript works with
them.
Added error message div on html form for jquery to populate.
Enabled javascript for tests that expect query params.
Removed no_js_selenium driver as it wasn't working as expected (js was
not actually off).

Authors: @lloydnye @oswaldquek @yolinas