bau: Improve styling on intermediate size devices
Between mobile and desktop sizes the radio buttons were wrapping. We
have introducted a min-width like the old frontend to make them look
correct. It is currently a hardcoded em value, so could be improved.

We have only applied the min-width to the checkbox on the select
documents page. Applying it to all block labels causes css issues with
the sign-in/registration highlighting.

Authors: @yolinas @lloydnye @oswaldquek @richardtowers @tombye