BAU: Remove duplicated code from IDP eligibility checker
The IdpEligibility::Checker class defined a mask consisting of all the
document evidence. This was duplicated in Evidence::DOCUMENT_ATTRIBUTES
which would mean two places to maintain a list of document-related
evidence items. Evidence has now been made a submodule of
IdpEligibility which had the added bonus of removing some requires from
specs.

Author: @vixus0