fix: logstash logger was breaking raven logger
The logstash logger was breaking the expected behaviour of the rails logger
which allows for delegation to extra loggers. This was preventing us from
logging to raven when erroring. To fix this we stopped using the LogStashLogger
rails config and created a logger explicitly that is added to the Rails logger
as a new delegate (just like we do with Raven).

Authors: @chrisholmes @lloydnye