bau: Fix checkbox focus / blur in Internet Explorer
Rails outputs two inputs for a checkbox so that form posts always
contain that field, even when the user has not checked the checkbox.
We ended up with HTML like this:

```
label.block-label
  input[type=hidden]
  input[type=checkbox]
```

Apparently having a label wrapping two inputs confuses IE into not
triggering the focus / blur events of the inputs when the label is
focused / blurred.

We've fixed this by using Rails' label helper to give the label an id.

```
label.block-label[for=no_documents]
  input[type=hidden]
  input[type=checkbox]#no_documents
```

Pair: @richardtowers @oswaldquek