3568: start storing evidence answers in session cookie
Stores the selected evidence answers in the session, but for now they will not
be read anywhere in the application.

The intention is that we will use the session store to remove the use of the
propagated query params.

Answers are currently stored and associated with a key for the type of
evidence/page they are related to. This way we should be given a guarantee that
only the most recent answers to questions are stored in the cache.