bau: Add old `-cy` routes back in to avoid downtime
If a user has loaded a `-cy` route when the deployment happens
they could end up submitting it to a version of front end that doesn't
know about that route anymore. This commit adds support for the old
routes in alongside the new routes to avoid this problem.

Solo: @richardTowers