3559: Use Puma for Capybara tests
This pulls in the capybara-puma gem, which configures Capybara to use
Puma instead of WEBrick when spinning up the app.

This is to support the fingerprint JS test: if the browser being used has
lots of plugins, the fingerprint request can be too large for WEBrick
(possibly when it's over 2048 bytes) which causes the test to fail.

Authors: @richardtowers @leoshaw