3547: Set locale from journey hint more neatly
In the Authn Request / Response controllers the locale should be pulled
from the journey hint cookie instead of the request params. Previously
we couldn't work out how to set this globally, because once it's set
once you can't change it. By overriding set_locale we're able to set it
once to the correct value.

Authors: @richardTowers @veda @lloydnye