BAU: Remove unecessary javascript-enabled tests
Lots of these tests were written when we were still using query
parameters to pass selected evidence along. The tests had to be run in
a real browser for these parameters to be preserved. As we are now
using the session store for selected evidence, the tests can run
without a browser.

Authors: @vixus0 @thomaslanlee