3626: announce form changes when selecting no documents checkbox
Sighted users can see the checkboxes change to ‘no’ when the checkbox on the select documents form is checked. To replicate this behaviour for screenreader users we need to announce this change dynamically.

A visually hidden message is read out at the point that the checkbox is activated. When it is deactivated the message is blanked, but that doesn’t get read out because it’s empty.

Author: @robinwhittleton