3642: Ensure new frontend release is ZDD compliant
The old frontend has two endpoints, /other-ways-to-access-service and
/other-ways-to-access, that both show the same thing. See
https://github.gds/gds/ida-frontend/blob/master/frontend/src/main/java/uk/gov/ida/hub/frontend/Urls.java#L141.

The difference between them is that /other-ways-to-access is not accessible if
the user is not in a AuthnFailedErrorState. Agreed with @chrisholmes that we
don't care about this anymore as the user can see this information even if not
in an AuthnFailedErrorState.

However to ensure the release of this story is ZDD compliant this commit adds a
/other-ways-to-access route which will be removed at a later point.

@oswaldquek