3628: store current identity providers in session
Identity Providers will be lazily loaded into the session object as they're
required. This should reduce the number of internal API calls needed when we
need to work with the list of available IDPs.

This has been taken as a step towards returning the list of IDPs at the start of
a session alongside the session's transaction.

I'm of two minds regarding whether this code should be placed inside a mixin or
interstitial controller base class (e.g. ChooseACertifiedCompanyController
< FederationController < ApplicationController) so I'm open to opinions.