bau: don't send raven log messages to sentry
If raven-ruby encounters an error while sending a message to Sentry (e.g.
a network error) it will use the rails logger to log an error message.
Unfortunately, we try to log error messages from the Rails logger to Sentry
using Raven so this introduces a bit of a loop which can result in a segfault if
the network error persists.

This change instructs our mechanism that logs messages to sentry from the Rails
logger to ignore messages that start with the Raven log message prefix.