3731: Report IdP selection/history in same page view
In #3492 we repurposed a custom variable to record the user's IdP
selection history at the point they select a new IdP for registration.
We were using a new virtual page view to set the custom variable, but
this was causing problems for the analytics team because they were
relying on the transition from the previous virtual page view to the
user returning from the IdP.

This change combines the setting of the IdP history custom variable with
the existing 'REGISTER_IDP' one so we're only making one page view to
set both custom variables.

Authors: @yolinas @leoshaw