3747: select documents legends for screenreaders
Previously we’ve used spans instead of legends for the headings on the select documents pages. legends are difficult to style in general and we can’t make them display in the way we want. We would like the accessibility benefits that legends bring for screenreaders though. @accessiblewebuk suggested that we could have both: a visually hidden legend and an audibly hidden span.

Authors: @robinwhittleton @accessiblewebuk