BAU: Fix feedback character counter in IE9
IE9 advertises support for the input event, but has bugs with backspace / deletion and pasting text. As the keydown event is currently removed if input is supported, this means IE9 will never update the character account if a backspace is used, and will permanently be off from the correct amount from that point.

We can integrate a jQuery input event normalisation plugin to fix this. This also has the side benefit of fixing IE8’s event to allow for updates during a held down keypress, something it previously couldn’t do with the keydown event it was using.

Author: @robinwhittleton