3733: First release with support for new flow
This adds support for a flag in the session which if set will show the
'Age' questions first.  By default it isn't set - the next release will
make it the default, but it's necessary to do it in two steps to avoid
breaking journeys that are in flight at the time of the roll-out.

Authors: @adityapahuja @leoshaw