BAU: upgrade jQuery Validate library
Currently we use a 3rd-party gem that seems to be mostly unsupported. jQuery Validate is two minor versions ahead of the gem, and the gem as it stands would block us upgrading to Rails 5. Given that, this PR removes the gem and direction includes a copy of the library.

There are numerous updates between 1.13.1 and 1.15.1: https://github.com/jzaefferer/jquery-validation/compare/1.13.1...1.15.0 . Most of them seem to be general code maintenance, better code localisation and small bug fixes in parts of the system we don’t use.

Author: @robinwhittleton