bau: reset locale to english before examples
When hitting a path we set the locale using I18n to the correct locale for that
path (:en or :cy), but because all our tests run in the same thread the locale
was kept until we hit a different locale path in another example. As the _path
helper uses the locale to choose an appropriate path we were seeing tests hit
welsh routes when we were intending for english ones if the the example followed
a welsh language example.