3795: use custom radio partial on start page
The block label partial requires a form model to generate the inputs. This switches the start form over to use a simple model with the same validation rules.

The non-JS error message has been moved above the form fields to match the rest of the app. JS error messages remain below.

Author: @robinwhittleton