3580: log API server errors at warn
If an error/problem occurs upstream within the Hub we receive a formatted
error message. As our upstream services also log their errors it shouldn't
be necessary to log them at a high level at the FE so I have changed it to only
log if it sees a message in an unexpected format (missing a type field).

Connection and Timeout errors will still be logged at error.