refactor: use rp display data for reporting analytics
Rather than calling the federation translator directly and having to encode
knowledge of the locales inside the analytics reporter we can pass in the
RpDisplayData object that is available for our transaction.

In addition, we now check the locale files for the RP's analytics description
on start up of the app so we shouldn't get into a state where it is missing.