ZD22311: Show validation errors even when js-hidden
Some users were seeing problems with the select-phone page where the JS
to show the conditional questions would not work. This left them in a
state where the form was invalid, but they were unable to fill in the
mandatory questions.

This ensures that we always show hidden questions when they are marked
as invalid.

Authors: @richardtowers @andrien.ricketts