stub-api: make /test-saml fall back to correct relay state
The saml-response-post button on the /test-saml page allows you to
simulate the response we receive from an IdP once a user has been
succesfully identified.

The relay state sent back should match the user's session id. For
automated test purposes we can pass this as a query param. When running
with a stubbed API it's convenient to be able to navigate straight to
/test-saml and use it as an entry point to the front end. When using it
like this it's convenient if it sets the RelayState correctly.

Solo: @richardtowers