stub-api: Add sinatra app to stub closed source APIs
Also add optional `--stub-api` flag to `./startup.sh` and document this
as the primary way of starting the front end. This will mean that
developers elsewhere in GDS and elsewhere in the world will be able to
play with the open source front end with a minimum of faff.

TODO: tests

Solo: @richardtowers