stub-api: Add unit tests for stub_api
These check that the reponses from the various stub_api endpoints are
valid according to the DTOs used by SessionProxy / TransactionsProxy.
This should ensure that if we change the format of the APIs we're
reminded to change the canned responses that stub_api returns.

It would be nice to have some feature tests to exercise this too, but I
can't think of a neat way of bringing the application with API_HOST
pointing at the stub_api server. We could naively point at localhost:50190, but
that would mean you had to stop the old front end to free up the port
before you could run the tests. :sick:

It may be a better idea to have stub_api.rb use the DTOs instead of
hardcoding strings. This would mean the tests could just call the
endpoints and check the status codes. It somewhat ruins the minimalist
aesthetic of the file though, so I haven't bothered.