stub-api: Add idp-authn-response endpoint
This allows you to (sort of) test the IdP return journey. You need to go
through the normal journey first to get a selected-idp in your session,
then you can go to test-saml and choose saml-response-post. You'll end
up on the /failed-sign-in path.

Maybe stubbing this to provide a success response would be better, but
good enough for now.