3342: set a cookie to enable routing to new cycle 3 page
We will be using a temporary cookie to enable routing through to the new cycle
three page. It will be set on all new journeys and users will be routed via
nginx to the new cycle 3 page only if the cookie is set correctly.

After 2 hours we will remove the cookie routing on nginx followed by the removal
of the cookie.