bau: Fix a test that seems to depend on resolution
This tests was failing consistently for me when firefox was popping up
on my laptop screen. I had a play around with screen resolutions and it
looks like the test fails on low resolutions and passes on high
resolutions. I think this is due to a bug (?) in Capybara / Selenium
where it can't click on an element that is "below the fold".

If we set and explicit size the firefox window this should avoid the issue
in most cases. A better fix would be to do some more investigation and fix
this upstream in capybara - volunteers welcome!

Solo: @richardtowers