5017: Dealing with multiple experiments so that they can be tidied and introduced easily
As we can have multiple values in the cookie we are storing a hash. We serialize to JSON and deserialize when reading.
Feature level tests use rails cookie handling so they need to CGI.escape cookies when setting them in order to be valid.
This could be added to the set cookie helper methods.
Non feature tests that access the cookie hash directly don't need this escaping as this is how the values will be available when rails is creating the hash at runtime.

Authors: @lloydnye @oswalkquek @datamineruk