bau: store session start time as integer
We're running into ZDD problems as the current start time in production is an
integer, but we were introducing a timestamp replacement which the currently
deployed FE didn't know what to do with.

If we want to introduce a timestamp we should raise this as a separate story.

Authors: @chrisholmes @willp-bl