bau: log session cookie size if it's large
We store a lot of stuff in the session cookie. This commits add code
to log a warning if it's getting too large.  Without going too deep
into Rails we can log an approximation of the size, at a slightly
lower threshold.

Perhaps we can move this into the ApplicationController; I don't mind
too much as long as we can have some sort of visibility of this
becoming an issue.

Pair: WillPa