5072: report metrics around HTTP status code counts
We would like to see the occurrences of different status codes in our responses
to users so that we can provide a visualisation relating to the health of our
system (count of 2xxs vs 3xxs vs 4xxs vs 5xxs).

I have used the statsd/notifications metric reporter system to provide these metrics

/service-status is currenty hard coded as testing with service_status_path or
dependency injecting it doesn't appear to be possible.