BAU: Remove use of have_not_been_made
have_not_been_made does not match requests as expected. It requires the very specific request to have been made in order to fail.
The actual page here does make a request to Piwik (because the user hits the sign in page), so I have changed the test to show that.

Solo: @lloydnye