5057: Add number of idps to idp selected piwik report
We're reporting "index x of n" so that we can see if users prefer to
click IdPs that are shown at the top of the list. We need to report n as
well as x because depending on the evidence the user said they had we
may not always show them the same number of IdPs. This means that users
will click index 4 less often, because they're often not even shown that
many IdPs. By storing the total we can account for this in our query.

Authors: @richardtowers @adityapahuja