5057: Refactor feature tests check for registration piwik report
When the user registers with an IdP we report the position the IdP was
in in the list when they clicked on it. The report now has the form:

"#{idp_name} was chosen for registration \(#{recommended_str}\) \(index
 #{idp_index} of #{idp_count}\) with evidence #{evidence}"

This commit forces you to specify how many IdPs there were and which IdP
was clicked when stubbing this report.

In some cases the tests don't click on an IdP, which means that the
selected IdP + count won't be in flash. In this case (index - of -) is
reported.

In all the other cases the tests pick an IdP by name. Because the IdP
list is shuffled, this means the selected IdP position is random. To
work around this we're specifying "\d" as the IdP index (meaning "any
digit").

Authors: @richardtowers @adityapahuja