5057: Add flash.keep to RedirectToIdpWarningController#index
We're setting two values in flash in the ChooseACertifiedCompanyController.
The [docs](http://api.rubyonrails.org/classes/ActionDispatch/Flash.html)
say that flash is "exposed to the very next action and then cleared
out". This would seem to suggest that the values we stored would be
available in RedirectToIdpWarningController#index, but not in #continue
or #continue_ajax where we read them. This doesn't agree with manual
testing however, it seems that the values /are/ still in flash when the
page is submitted.

In the interests of following the docs, we've added flash.keep() to the
index method. This means that Rails will not discard the flash until
after the next action (the form post), so they're guaranteed to be
available when we read them.

Note that there's still an edge case where the user proceeds to the IdP
and then presses back. In this case their flash /will/ have been
cleared, and we'll report (index - of -) instead of (index 1 of 3). This
is a compromise we're willing to make.

Authors: @richardtowers @adityapahuja