3429: Set a simple, lenient CSP header
Content-Security-Policy (CSP) is applied to restrict browser behavior to
the features used within the verify-frontend.

This policy is purposefully quite lenient, allowing any inline-scripts
to execute. We intend to add a reporting-uri in the future and iterate
our policy to further restrict browser behavior.

Specifically this policy only allows resources to be fetched from
'self' (origin of the document being served) it further restricts the
fonts to load using the 'data' scheme as they are provided from the
frontend_toolkit. We disallow embedding any object/applets. Scripts can
be loaded from self or as inline html, the same policy is applied to
stylesheets.