BAU: layout bug on will-it-work-for-me
The reveal panel has a class of “inline” even though the options should be stacked. This is because each of the text in the options is long enough to force the next option to the next line. Removing the class will cause them to naturally stack regardless of the text.

Author: @robinwhittleton