bau: Switch to use jasmine gem
We've switched primarily because jasmine-rails was producing a lot of
deprecation warnings on rails 5. The jasmine gem is maintained by the jasmine team
so this is probably the better long term option for updates/upgrades.

NOTE: we had to use document.onsubmit instead of $(document).on('submit') for
some formspy code as this old implementation was broken with this gem.