Use new GOV.UK-wide radios and checkboxes
This updates GOV.UK Elements to 2.2.0 which includes the new GOV.UK-wide radios and checkboxes. It also contains some minor style fixes around the integration.

In the new scheme the native radios and checkboxes are hidden with opacity:0; which breaks Capybara’s `choose` function. We can click on the label instead, but that required an update to Capybara. There are deprecation warnings, but can be fixed in a future commit.