Override default elements radio colours on start
The default radio button colour matches the text colour and background that it sits in. On Verify’s start page this is a bit of a change compared to the existing style. This block of style overrides the colours to make it look more like existing Verify.

It would be interesting to A/B test these Verify styles against the elements defaults, as they have benefits for users who want to pick their own colours.