5482: AB test for IDP rankings
Adds a ranking step to the Choose a Certified Company page. The ranking
is defined in verify-frontend-federation-config.

NOTE: we are still waiting for Olly's ranking data. Therefore we are
not sending requests to piwik!

We have test data in place to test it only in the Joint
environment.

Authors: @srinivasmurty @phss