TT-161: Create photo documents page for an ab test
A new ab test is about splitting out questions regarding cases where the
user has or has not documents with a photo in it. This commit adds the
first page of that split

We squashed 7 WIP commits together into this commit to avoid confusing
git history.

Authors: @DataMinerUK @nasashrafthoughtworks @AndrienRicketts