bau: hotfix for 404 Sentry Message Spam
404 messages (and all exceptions) are currently being logged in
[DebugExceptions](https://github.com/rails/rails/blob/master/actionpack/lib/action_dispatch/middleware/debug_exceptions.rb#L169)
across multiple log messages so we need to do some filtering here to
ignore each message to reduce the unactionable messages we send to
Sentry.

A fix was made in Rails 5.0.2 to not log the stacktrace lines using
multiple messages, but the Exception class and message and the
stacktrace are still being logged as separate messages. We need to
raise an issue with Rails to solve this problem

The problems in Rails were introduced in the following two commits:
* https://github.com/rails/rails/commit/632938cfda8762a98a7c7a360a6f05c29d2cdb5e#diff-8bdcea77b8b971917390352c57a58653
* https://github.com/rails/rails/commit/89fab56597c335bb49887563b9a98386b5171574#diff-8bdcea77b8b971917390352c57a58653

This behaviour was introduced into Rails to support displaying tags on
each line in the error and stacktrace message, but this should really be
implemented at the Logger implementation level.