Move to new elements form error classes.
Elements 3.0 introduces new error classes to allow for more fine-grained control over visualisation of error states. None of this particularly affects Verify, so our work to update is basically just to change `error` to `form-group-error` on things that already have `form-group`, and to change the error class applied directly to inputs to be `form-control-error`.