TT-227: Order Verify servies by volume of use
This commit makes it such that the RP's listed in relying_parties.yml's
transaction_type.display_name_and_homepage property are listed in that
order when a list of services is shown on the session timeout page, and
the 'about identity accounts' page.

A test in display_data_correlator_spec.rb that throws an error when a simpleId
json field is missing from the transaction data returned by the config service
API has been removed as we think this shouldn't stop the list of services from
being shown on the pages. If a simpleId field json field is missing, the code
deals with that and doesn't shown that RP.

@oswaldquek @sgreensmith