bau: add COOKIE_NOT_FOUND_PAGE page source
https://github.com/alphagov/verify-frontend/blob/7379b463065c0d9f52413daa012c3fbc2e1ea2f8/spec/features/user_visits_start_page_spec.rb#L38
shows that there is a COOKIE_NOT_FOUND_PAGE page source. This commit adds it to
the feedback_source_mapper.rb. It doesn't map to any url as no cookies means no
session, so the 'Your session has timed out' text should show after the
submission of feedback. This is reflected in the new test added in
user_visits_feedback_page_spec.rb.

@oswaldquek