TT-225:  Added ability to show whether an IDP requires mobile app installation
First, we have added new A/B support for routes that will use A and B
versions of the page that shows IDPs (the choose_a_certified_company
page).  Next, we have added a new property to the IDP YML file, which
will contain the message showing whether an IDP will or may need the
user to install a mobile application.  To support the current AB test,
we modified the entry in ab_test.yml to force the journey to use the B
route.

Authors: @naseerashraf, @kgarwood